Add default client to MCPS audience
